+++ applications/fishing-map/src/features/app/app.selectors.ts
@@ -135,7 +135,7 @@ export const selectCustomWorkspace = createSelector(
   ): WorkspaceUpsert<WorkspaceState> => {
     return {
       ...workspace,
-      name: `${user?.id}_${Date.now()}`,
+      name: `user_${user?.id}_${Date.now()}`,
       app: APP_NAME,
       aoi: undefined,
       dataviews: workspace?.dataviews?.map(({ id }) => id as number),