+++ applications/fishing-map/src/features/datasets/datasets.slice.ts
@@ -1,18 +1,32 @@
 import { createAsyncThunk, createSelector } from '@reduxjs/toolkit'
 import memoize from 'lodash/memoize'
+import uniqBy from 'lodash/uniqBy'
+import uniq from 'lodash/uniq'
 import { Dataset } from '@globalfishingwatch/api-types'
 import GFWAPI from '@globalfishingwatch/api-client'
 import { AsyncReducer, createAsyncSlice } from 'utils/async-slice'
 import { RootState } from 'store'
+
+const DATASETS_CACHE = true
 
 export const fetchDatasetsByIdsThunk = createAsyncThunk(
   'datasets/fetch',
   async (ids: string[], { rejectWithValue }) => {
     // TODO fetch only not already existing ids
     try {
-      let datasets = await GFWAPI.fetch<Dataset[]>(
-        `/v1/datasets?ids=${ids.join(',')}&include=endpoints&cache=false`
+      const initialDatasets = await GFWAPI.fetch<Dataset[]>(
+        `/v1/datasets?ids=${ids.join(',')}&include=endpoints&cache=${DATASETS_CACHE}`
       )
+
+      const relatedDatasetsIds = initialDatasets.flatMap(
+        (dataset) => dataset.relatedDatasets?.flatMap(({ id }) => id || []) || []
+      )
+      const uniqRelatedDatasetsIds = uniq([...relatedDatasetsIds, ids]).join(',')
+      const relatedDatasets = await GFWAPI.fetch<Dataset[]>(
+        `/v1/datasets?ids=${uniqRelatedDatasetsIds}&include=endpoints&cache=${DATASETS_CACHE}`
+      )
+
+      let datasets = uniqBy([...initialDatasets, ...relatedDatasets], 'id')
       if (process.env.REACT_APP_USE_DATASETS_MOCK === 'true') {
         const mockedDatasets = await import('./datasets.mock')
         datasets = [...datasets, ...mockedDatasets.default]
@@ -23,11 +37,14 @@ export const fetchDatasetsByIdsThunk = createAsyncThunk(
     }
   }
 )
+
 export const fetchDatasetsByIdThunk = createAsyncThunk(
   'datasets/fetchById',
   async (id: string, { rejectWithValue }) => {
     try {
-      const dataset = await GFWAPI.fetch<Dataset>(`/v1/datasets/${id}?include=endpoints`)
+      const dataset = await GFWAPI.fetch<Dataset>(
+        `/v1/datasets/${id}?include=endpoints&cahe=${DATASETS_CACHE}`
+      )
       return dataset
     } catch (e) {
       return rejectWithValue(id)