+++ applications/fishing-map/src/features/dataviews/dataviews.slice.ts
@@ -4,21 +4,15 @@ import { Dataview } from '@globalfishingwatch/api-types'
 import GFWAPI from '@globalfishingwatch/api-client'
 import { AsyncReducer, createAsyncSlice } from 'utils/async-slice'
 import { RootState } from 'store'
-import { getDatasetByDataview } from 'features/workspace/workspace.slice'
-import { fetchDatasetsByIdsThunk } from 'features/datasets/datasets.slice'
 
 export const fetchDataviewsByIdsThunk = createAsyncThunk(
   'dataviews/fetch',
-  async (ids: number[], { dispatch, rejectWithValue }) => {
+  async (ids: number[], { rejectWithValue }) => {
     try {
       let dataviews = await GFWAPI.fetch<Dataview[]>(`/v1/dataviews?ids=${ids.join(',')}`)
       if (process.env.REACT_APP_USE_DATAVIEWS_MOCK === 'true') {
         const mockedDataviews = await import('./dataviews.mock')
         dataviews = [...dataviews, ...mockedDataviews.default]
-      }
-      const datasets = getDatasetByDataview(dataviews)
-      if (datasets?.length) {
-        await dispatch(fetchDatasetsByIdsThunk(datasets))
       }
       return dataviews
     } catch (e) {