+++ applications/fishing-map/src/features/search/search.slice.ts
@@ -4,10 +4,15 @@ import { resolveEndpoint } from '@globalfishingwatch/dataviews-client'
 import { Dataset, Vessel, APISearch } from '@globalfishingwatch/api-types'
 import { RootState } from 'store'
 import { AsyncReducerStatus } from 'types'
+import { selectDatasetById } from 'features/datasets/datasets.slice'
+import { getRelatedDatasetByType } from 'features/workspace/workspace.selectors'
+import { TRACKS_DATASET_TYPE } from 'data/datasets'
+
+export type VesselWithDatasets = Vessel & { dataset: string; trackDatasetId?: string }
 
 interface SearchState {
   status: AsyncReducerStatus
-  data: APISearch<Vessel> | null
+  data: VesselWithDatasets[] | null
 }
 
 const initialState: SearchState = {
@@ -17,7 +22,8 @@ const initialState: SearchState = {
 
 export const fetchVesselSearchThunk = createAsyncThunk(
   'search/fetch',
-  async ({ query, datasets }: { query: string; datasets: Dataset[] }) => {
+  async ({ query, datasets }: { query: string; datasets: Dataset[] }, { getState }) => {
+    const state = getState() as RootState
     const dataset = datasets[0]
     const datasetConfig = {
       endpoint: 'carriers-search-vessels',
@@ -30,8 +36,25 @@ export const fetchVesselSearchThunk = createAsyncThunk(
     }
     const url = resolveEndpoint(dataset, datasetConfig)
     if (url) {
-      const searchResults = await GFWAPI.fetch<any>(url)
-      return searchResults
+      const searchResults = await GFWAPI.fetch<APISearch<Vessel>>(url)
+      const largerResult = Math.max(...searchResults.map(({ results }) => results.limit))
+      const resultsFlat = Array.from(Array(largerResult).keys()).flatMap((index) => {
+        // Flat them in order of results so for examplen when requesting 3 datasets the list will be
+        // dataset1[0], dataset2[0], dataset3[0], dataset1[1], dataset2[1], dataset3[1]
+        return searchResults.flatMap(({ dataset, results }) => {
+          const vessel = results.entries[index]
+          if (!vessel) return []
+
+          const infoDataset = selectDatasetById(dataset)(state)
+          const trackDatasetId = getRelatedDatasetByType(infoDataset, TRACKS_DATASET_TYPE)?.id
+          return {
+            ...vessel,
+            dataset,
+            trackDatasetId,
+          }
+        })
+      })
+      return resultsFlat
     }
   }
 )
@@ -50,7 +73,9 @@ const searchSlice = createSlice({
     })
     builder.addCase(fetchVesselSearchThunk.fulfilled, (state, action) => {
       state.status = AsyncReducerStatus.Finished
-      state.data = action.payload
+      if (action.payload) {
+        state.data = action.payload
+      }
     })
     builder.addCase(fetchVesselSearchThunk.rejected, (state) => {
       state.status = AsyncReducerStatus.Error