+++ applications/fishing-map/src/features/sidebar/vessels/VesselLayerPanel.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react'
+import React, { Fragment, useCallback, useState } from 'react'
 import cx from 'classnames'
 import { useSelector } from 'react-redux'
 import { useTranslation } from 'react-i18next'
@@ -27,12 +27,14 @@ import I18nFlag from 'features/i18n/i18nFlag'
 import { useMapboxInstance } from 'features/map/map.context'
 import useViewport from 'features/map/map-viewport.hooks'
 import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
+import { DATAVIEW_INSTANCE_PREFIX } from 'features/dataviews/dataviews.utils'
 
 // Translations by feature.unit static keys
 // t('vessel.flag', 'Flag')
 // t('vessel.imo', 'IMO')
 // t('vessel.first_transmission_date', 'First transmission date')
 // t('vessel.last_transmission_date', 'Last transmission date')
+// t('vessel.fleet', 'Fleet')
 
 type LayerPanelProps = {
   dataview: UrlDataviewInstance
@@ -96,10 +98,6 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
     deleteDataviewInstance(dataview.id)
   }
 
-  const datasetConfig = dataview.datasetsConfig?.find(
-    (dc: any) => dc?.params.find((p: any) => p.id === 'vesselId')?.value
-  )
-
   const vesselName = resource?.data?.shipname
 
   const onToggleColorOpen = () => {
@@ -115,7 +113,7 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
   }
   const expandedContainerRef = useClickedOutside(closeExpandedContainer)
 
-  const vesselId = datasetConfig?.params.find((p: any) => p.id === 'vesselId')?.value as string
+  const vesselId = dataview.id.replace(DATAVIEW_INSTANCE_PREFIX, '')
   const title = vesselName || vesselId || dataview.name
 
   const TitleComponent = (
@@ -123,6 +121,10 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
       {title && formatInfoField(title, 'name')}
     </h3>
   )
+
+  const loading =
+    trackResource?.status === AsyncReducerStatus.Loading ||
+    resource?.status === AsyncReducerStatus.Loading
 
   return (
     <div
@@ -142,46 +144,58 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
           TitleComponent
         )}
         <div className={cx(styles.actions, { [styles.active]: layerActive })}>
-          <IconButton
-            icon="target"
-            size="small"
-            className={styles.actionButton}
-            tooltip={t('layer.vessel_fit_bounds', 'Center view on vessel track')}
-            onClick={onFitBoundsClick}
-            tooltipPlacement="top"
-          />
-          <IconButton
-            icon="info"
-            size="small"
-            loading={resource?.status === AsyncReducerStatus.Loading}
-            className={styles.actionButton}
-            tooltip={
-              infoOpen ? t('layer.info_close', 'Hide info') : t('layer.info_open', 'Show info')
-            }
-            onClick={onToggleInfoOpen}
-            tooltipPlacement="top"
-          />
-          {layerActive && (
+          {loading ? (
             <IconButton
-              icon={colorOpen ? 'color-picker' : 'color-picker-filled'}
+              loading
+              className={styles.loadingIcon}
               size="small"
-              style={colorOpen ? {} : { color: dataview.config?.color }}
-              tooltip={t('layer.color_change', 'Change color')}
-              tooltipPlacement="top"
-              onClick={onToggleColorOpen}
-              className={cx(styles.actionButton, styles.expandable, {
-                [styles.expanded]: colorOpen,
-              })}
+              tooltip={t('vessel.loading', 'Loading vessel track')}
             />
+          ) : (
+            <Fragment>
+              {layerActive && (
+                <Fragment>
+                  <IconButton
+                    icon={colorOpen ? 'color-picker' : 'color-picker-filled'}
+                    size="small"
+                    style={colorOpen ? {} : { color: dataview.config?.color }}
+                    tooltip={t('layer.color_change', 'Change color')}
+                    tooltipPlacement="top"
+                    onClick={onToggleColorOpen}
+                    className={cx(styles.actionButton, styles.expandable, {
+                      [styles.expanded]: colorOpen,
+                    })}
+                  />
+                  <IconButton
+                    icon="target"
+                    size="small"
+                    className={styles.actionButton}
+                    tooltip={t('layer.vessel_fit_bounds', 'Center view on vessel track')}
+                    onClick={onFitBoundsClick}
+                    tooltipPlacement="top"
+                  />
+                </Fragment>
+              )}
+              <IconButton
+                icon="info"
+                size="small"
+                className={styles.actionButton}
+                tooltip={
+                  infoOpen ? t('layer.info_close', 'Hide info') : t('layer.info_open', 'Show info')
+                }
+                onClick={onToggleInfoOpen}
+                tooltipPlacement="top"
+              />
+              <IconButton
+                icon="delete"
+                size="small"
+                className={styles.actionButton}
+                tooltip={t('layer.remove', 'Remove layer')}
+                tooltipPlacement="top"
+                onClick={onRemoveLayerClick}
+              />
+            </Fragment>
           )}
-          <IconButton
-            icon="delete"
-            size="small"
-            className={styles.actionButton}
-            tooltip={t('layer.remove', 'Remove layer')}
-            tooltipPlacement="top"
-            onClick={onRemoveLayerClick}
-          />
         </div>
       </div>
       <div className={styles.expandedContainer} ref={expandedContainerRef}>